loader/requirements.txt
mysql-connector-python==8.0.23
pandas==1.2.2
numpy==1.19.2